# aval/contracts/api.yaml
openapi: 3.1.0
info:
  title: Aval Contracts (api + merchant)
  version: 1.1.0
  description: |
    Contrato de APIs para construcción paralela (M0). Dos servicios:
    - `api` (Canal de Confianza): mandates [A], verify/purchases/audit/events [B], escalations/bot [A]
    - `merchant` (VuelaYa): catálogo + checkout + webhooks del rail Yuno-style simulado [D]
    Reglas: cambios aditivos = v1.x (PR + mock + trustlib en el mismo commit). Cambios rompibles solo antes de M2.
servers:
  - url: https://api.{domain}
    variables:
      domain: { default: localhost, description: dominio propio del demo }
  - url: https://merchant.{domain}
    variables:
      domain: { default: localhost }
...
paths:
...
  /escalations:
    get:
      tags: [escalations]
      summary: Escalamientos (filtro por mandate/estado)
      parameters:
        - { name: mandate_id, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [pending, resolved, expired] } }
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Escalation" }
  /escalations/{id}/resolve:
    post:
      tags: [escalations]
      summary: |
        Resolver escalamiento (bot Telegram o UI). APPROVE con flag sticky=true
        crea mini-mandato derivado firmado. Timeout 120 s → auto-deny (fail-closed).
      parameters: [ { $ref: "#/components/parameters/IdPath" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResolveRequest" }
      responses:
        "200":
          description: Resuelto (+ resume de la saga si APPROVE)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Escalation" }
        "409": { $ref: "#/components/responses/Rejection" }
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
...
    Rejection:
      description: Rechazo con reason_code
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Rejection" }
  schemas:
    ReasonCode:
      type: string
      enum:
        - AMOUNT_EXCEEDS_PER_TXN
        - BUDGET_EXCEEDED
        - LIMIT_EXHAUSTED
        - CATEGORY_FORBIDDEN
        - MERCHANT_NOT_ALLOWED
        - MANDATE_EXPIRED
        - MANDATE_NOT_YET_VALID
        - MANDATE_REVOKED
        - MANDATE_SUSPENDED
        - MANDATE_EXHAUSTED
        - CONDITION_FAILED
        - INVALID_SIGNATURE
        - INVALID_PROOF_OF_POSSESSION
        - DUPLICATE_JTI
        - NONCE_REUSED
        - ESCALATION_TIMEOUT_DENIED
        - RAIL_ERROR
        - RAIL_TOKEN_DELETED
...
    Escalation:
      type: object
      properties:
        escalation_id: { type: string }
        mandate_id: { type: string }
        purchase_id: { type: string }
        status: { type: string, enum: [pending, resolved, expired] }
        diff: { type: object }
        timeout_at: { type: string, format: date-time }
        resolution:
          type: object
          properties:
            decision: { type: string, enum: [APPROVE, REJECT] }
            approver: { type: string }
            channel: { type: string, enum: [telegram, web] }
            resolved_at: { type: string, format: date-time }
            receipt_sig: { type: string, description: approval receipt firmado (evidence bundle) }
    ResolveRequest:
      type: object
      required: [decision, approver, channel]
      properties:
        decision: { type: string, enum: [APPROVE, REJECT] }
        approver: { type: string }
        channel: { type: string, enum: [telegram, web] }
        sticky:
          type: object
          description: si viene, crea mini-mandato derivado
          properties:
            max_per_txn: { type: number }
            count: { type: integer }
            period: { type: string }
...
    Rejection:
      type: object
      required: [reason_code]
      properties:
        reason_code: { $ref: "#/components/schemas/ReasonCode" }
        message: { type: string }
        purchase_id: { type: string }
...